[FEATURE] timeout for processing ocr can now be set over command line for workers(see -help)
